+++ next/CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Improved VTL editor. Errors are better handled.
 - Some variables names have been reverted to legacy ones in French (e.g. "Libellé").
 - Some visual improvements to Select component.